requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.36
pandas==2.2.3
openpyxl==3.1.5
PyYAML==6.0.2
rapidfuzz==3.10.1
openai==1.75.0
python-dotenv==1.0.1
Werkzeug==3.1.3
pytest==8.3.4
